Adds a 'generic' ND binning option
Adds a new format for binning definitions:

```yaml
Binning:
  Axes:
    - VarStr: "RecoNeutrinoEnergy"
      VarBins: [0.,  0.5,  1.,  1.25, 1.5, 1.75, 2., 2.25, 2.5, 2.75,
                3., 3.25, 3.5, 3.75, 4., 5., 6., 10.]
    - VarStr: "ERecQE"
      Uniform: [10, 0, 10]
    - VarStr: "EHadRec"
      Uniform: [10, 0, 10]
```

To the rest of MaCh3 this will appear like a 1D analysis, where the
analysis variable is the global bin number. When using this option,
the XVarStr is set to "global_bin_number", which must be implemented
by all SamplePDF sub classes that want to use this feature. We can
integrate this into Core more directly in the future, but for now,
I think this works.

'flow bins in the ND binning are ignored, the 1D binning runs from
[-0.5,nbins-0.5]. The 'default' ROOT 'flow bins exist on the 1D binning
and so the global bin finder method returns a value of -1 for a over/under
flow on any individual axis.